Fix: Se corrigio errores de importacion de prisma client en frontend, se movió a backend y solo se consume en front
